Rewrite IRC connection system
As the previous IRC library was not invented here and didn't support
rotor, I've switched distirc to use a custom IRC implementation based on
rotor. At a later date, I'll probably release this as a separate
project, but for now, it will stay attached to this project until it's
mature enough.