Implement basic client-core authentication
Clients now have to send an `Authenticate` message with a user ID and
password before they can do anything. As a side-effect, this means we
now fully support having multiple users.

Currently passwords are all stored and sent in plain text. We should
probably do something to improve the security in the future.

This closes #2